+++ src/unused-exports/app.ts
@@ -1,21 +1,17 @@
-import * as path from "path";
-import { makeContext } from "./context";
-import { getExports } from "./exports";
-import { getOnlyProjectImports } from "./importedFiles";
-import { getImports } from "./imports";
-import { getNotUsed, TNotUsed, sortNotUsedFn } from "./notUsed";
-import { getParsedFiles } from "./parsedFiles";
-import { buildRelations, TRelation } from "./relations";
-import { getSourceFiles } from "./sourceFiles";
-import { getOnlyUsefullFiles } from "./usefullFiles";
+import * as path from 'path';
+import { makeContext } from './context';
+import { getExports } from './exports';
+import { getOnlyProjectImports } from './importedFiles';
+import { getImports } from './imports';
+import { getNotUsed, TNotUsed, sortNotUsedFn } from './notUsed';
+import { getParsedFiles } from './parsedFiles';
+import { buildRelations, TRelation } from './relations';
+import { getSourceFiles } from './sourceFiles';
+import { getOnlyUsefullFiles } from './usefullFiles';
 
-const fixPath = (path: string, prefixLen: number): string =>
-  path.substr(prefixLen).replace(/\\/g, "/");
+const fixPath = (path: string, prefixLen: number): string => path.substr(prefixLen).replace(/\\/g, '/');
 
-const makePathRelativeToProject = (
-  relations: TRelation[],
-  absPathToPrj: string
-): void => {
+const makePathRelativeToProject = (relations: TRelation[], absPathToPrj: string): void => {
   const pathDelim = path.delimiter;
   const len = absPathToPrj.length + pathDelim.length;
   relations.forEach((r) => {