+++ src/unused-exports/exports.ts
@@ -1,39 +1,61 @@
-import { TTsParsed, TTsExport } from "./parsedFiles";
-import { TImport } from "./imports";
+import { TTsParsed, TTsExport, varNameRe } from './parsedFiles';
+import { TImport } from './imports';
 
-export const getExports = (
-  parsedFiles: TTsParsed[],
-  imports: TImport[]
-): TExport[] => {
+export const getExports = (parsedFiles: TTsParsed[], imports: TImport[]): TExport[] => {
   const arr: TExport[] = [];
   parsedFiles.forEach(parseExport(arr, imports));
   return arr;
 };
 
-const parseExport = (arr: TExport[], imports: TImport[]) => (
-  parsedFile: TTsParsed
-): void => {
+const parseExport = (arr: TExport[], imports: TImport[]) => (parsedFile: TTsParsed): void => {
   const { path, exports } = parsedFile;
   exports.forEach(addParsedExports(arr, imports, path));
 };
 
-const addParsedExports = (arr: TExport[], imports: TImport[], path: string) => (
-  anExport: TTsExport
-) => {
+const addParsedExports = (arr: TExport[], imports: TImport[], path: string) => (anExport: TTsExport) => {
   const { name, path: fromPath } = anExport;
-  const isUsed = itIsUsed(path, name, imports);
-  arr.push({
-    inPath: path,
-    name,
-    fromPath,
-    isUsed,
+
+  const names = getExportedNames(name);
+  names.forEach((name) => {
+    const isUsed = itIsUsed(path, name, imports);
+    arr.push({
+      inPath: path,
+      name,
+      fromPath,
+      isUsed,
+    });
   });
 };
 
+const groupRe = new RegExp(`${varNameRe}|,|\\{[^\\}]*\\}`, 'gi');
+const varNameInGroupRe = new RegExp(`${varNameRe}|,}`, 'gi');
+
+const getExportedNames = (name: string): string[] => {
+  const re = new RegExp(groupRe, 'gi');
+  const arr = [];
+  let res1: string[] | null;
+  while ((res1 = re.exec(name)) !== null) {
+    const foundName1 = res1[0];
+    const ch1 = foundName1[0];
+    if (ch1 !== '{') {
+      arr.push(foundName1);
+      continue;
+    }
+
+    const varNameInGroup = foundName1.substring(0, foundName1.length - 1);
+    let res2: string[] | null;
+    while ((res2 = varNameInGroupRe.exec(varNameInGroup)) !== null) {
+      const foundName2 = res2[0];
+      if (foundName2 !== ',') {
+        arr.push(foundName2);
+      }
+    }
+  }
+  return arr;
+};
+
 const itIsUsed = (path: string, name: string, imports: TImport[]): boolean =>
-  imports.find(
-    (anImport) => anImport.fromPath === path && anImport.name === name
-  ) !== undefined;
+  imports.find((anImport) => anImport.fromPath === path && anImport.name === name) !== undefined;
 
 export interface TExport {
   inPath: string;