+++ src/unused-exports/importedFiles.ts
@@ -1,32 +1,23 @@
-import * as glob from "glob";
-import * as path from "path";
-import * as vscode from "vscode";
-import { TContext } from "./context";
-import { isDirectory, isFile } from "./fsUtils";
-import { TTsImport, TTsParsed } from "./parsedFiles";
+import * as glob from 'glob';
+import * as path from 'path';
+import { TContext } from './context';
+import { isDirectory, isFile } from './fsUtils';
+import { TTsImport, TTsParsed } from './parsedFiles';
 
-export const getOnlyProjectImports = (
-  context: TContext,
-  parsedFiles: TTsParsed[]
-): TTsParsed[] => {
+export const getOnlyProjectImports = (context: TContext, parsedFiles: TTsParsed[]): TTsParsed[] => {
   const { baseUrl } = context;
 
   parsedFiles.forEach((tsParsed) => {
     const { path: filePath, imports } = tsParsed;
-    tsParsed.imports = imports
-      .map(makeImportAbs(baseUrl, path.dirname(filePath)))
-      .filter(importValid);
+    tsParsed.imports = imports.map(makeImportAbs(baseUrl, path.dirname(filePath))).filter(importValid);
   });
 
   return parsedFiles;
 };
 
-const importValid = (anImport: TTsImport): boolean =>
-  anImport.path !== undefined;
+const importValid = (anImport: TTsImport): boolean => anImport.path !== undefined;
 
-const makeImportAbs = (baseUrl: string | undefined, filePath: string) => (
-  anImport: TTsImport
-): TTsImport => {
+const makeImportAbs = (baseUrl: string | undefined, filePath: string) => (anImport: TTsImport): TTsImport => {
   const { path: relPath } = anImport;
 
   const absPath = path.resolve(filePath, relPath);
@@ -55,10 +46,7 @@ const makeImportAbs = (baseUrl: string | undefined, filePath: string) => (
   };
 };
 
-const resolveFilePath = (
-  filePath: string,
-  allowJs?: boolean
-): string | undefined => {
+const resolveFilePath = (filePath: string, allowJs?: boolean): string | undefined => {
   if (isFile(filePath)) {
     return filePath;
   }
@@ -66,7 +54,7 @@ const resolveFilePath = (
   /* try it as file */
   const globReFile = getGlobRegexp(filePath, allowJs);
   const resFile = glob.sync(globReFile, {
-    cwd: ".",
+    cwd: '.',
     nodir: false,
     nosort: true,
   });
@@ -81,7 +69,7 @@ const resolveFilePath = (
   /* try it as directory */
   const globReDir = getDirGlobRegexp(filePath, allowJs);
   const resDir = glob.sync(globReDir, {
-    cwd: ".",
+    cwd: '.',
     nodir: false,
     nosort: true,
   });
@@ -91,5 +79,4 @@ const resolveFilePath = (
 const getDirGlobRegexp = (path: string, allowJs?: boolean): string =>
   allowJs ? `${path}/index.(ts|js)?(x)` : `${path}/index.ts?(x)`;
 
-const getGlobRegexp = (path: string, allowJs?: boolean): string =>
-  allowJs ? `${path}.(ts|js)?(x)` : `${path}.ts?(x)`;
+const getGlobRegexp = (path: string, allowJs?: boolean): string => (allowJs ? `${path}.(ts|js)?(x)` : `${path}.ts?(x)`);